feat: add jwt validation and accommodation client communication

+++ user-service/service/user.service.go
@@ -18,15 +18,22 @@ type UserService struct {
 	validator         *utils.Validator
 	reservationClient *client.ReservationClient
 	authClient        *client.AuthClient
+	accClient         *client.AccClient
 }
 
-func NewUserService(userRepo *repository.UserRepository, jwtService *JwtService, validator *utils.Validator, reservationClient *client.ReservationClient, authClient *client.AuthClient) *UserService {
+func NewUserService(userRepo *repository.UserRepository,
+	jwtService *JwtService,
+	validator *utils.Validator,
+	reservationClient *client.ReservationClient,
+	authClient *client.AuthClient,
+	accClient *client.AccClient) *UserService {
 	return &UserService{
 		userRepository:    userRepo,
 		jwtService:        jwtService,
 		validator:         validator,
 		reservationClient: reservationClient,
 		authClient:        authClient,
+		accClient:         accClient,
 	}
 }
 
@@ -142,7 +149,7 @@ func (u *UserService) GetUserById(id string) (*domain.User, *errors.ErrorStruct)
 }
 
 func (u *UserService) DeleteUser(role string, id string) *errors.ErrorStruct {
-	if role != "Host" {
+	if role == "Guest" {
 		allow, err := u.reservationClient.GuestDeleteAllowed(context.TODO(), id)
 		if err != nil {
 			return err
@@ -153,6 +160,9 @@ func (u *UserService) DeleteUser(role string, id string) *errors.ErrorStruct {
 	}
 	if role == "Host" {
 		//TODO
+	}
+	if role != "Guest" && role != "Host" {
+		return errors.NewError("not allowed by role", 401)
 	}
 
 	newErr := u.authClient.DeleteUserAuth(context.TODO(), id)
